fixed show/remove document button in the restate edit window

+++ ui/src/main/java/com/re/components/util/MultiUpload.java
@@ -79,8 +79,9 @@ public class MultiUpload extends MultiFileUpload {
 
     private void addDocumentToDocementLayout(final REDocument reDocument) {
         RealEstate rEstate = reDocument.getRealEstate();
-        FileResource resource = rEstate != null ? new FileResource(reDocument.getDocument()): null;
-        final LoadedDocumentRow newRow = new LoadedDocumentRow(reDocument.getFileName(), resource);
+        final boolean isREstateExists = rEstate != null;
+        FileResource resource =  new FileResource(reDocument.getDocument());
+        final LoadedDocumentRow newRow = new LoadedDocumentRow(reDocument.getFileName(), resource, isREstateExists);
 
         loadedDocumentsLayout.addComponent(newRow);
         newRow.getDeleteButton().addClickListener(new Button.ClickListener() {
@@ -89,7 +90,8 @@ public class MultiUpload extends MultiFileUpload {
                 try {
                     loadedDocumentsLayout.removeComponent(newRow);
                     getReDocumentsList().remove(reDocument);
-                    reDocumentService.delete(reDocument);
+                    if (reDocument.getId() != null)
+                        reDocumentService.delete(reDocument);
                 } catch (IOException e) {
                     Notification.show("Не удалось удалить объект");
                 }
